chore: relocate collection schemas to proper folder
